[roll] Roll fuchsia [cmx] Delete tiles_ctl.
This tool relies on access to `/hub` which was removed from the
system a couple of months ago. Since the tool doesn't work on ToT
and we haven't heard from any developers regarding the breakage,
it seems safe to assume that it is no longer in use.

Original-Bug: 100732
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794976
Original-Revision: cdeb41a3e8ecf76f786f039d24ad2372ecdf18ad
GitOrigin-RevId: f1dbe393c4a1b6844c226f9bbfb3b37da414c736
Change-Id: I5984e9ce1f70e1b75a110dc23435fb30cc17240c