[roll] Roll fuchsia [netstack3] Access device addr status via context
Add a new trait for accessing IP per-device state and use it to
implement the existing address assignment state lookup. This is a no-op
for now but will make it possible to change all &SC arguments in context
and handler traits to &mut SC in a future CL.

This change is necessary to make with_devices take &mut self. Without
this, the impls of with_address_statuses will attempt to pass the same
&mut self into the closure that is an argument to with_devices, which
the compiler forbids.

Original-Bug: 120973
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796122
Original-Revision: 4871ad62ff498a6dc85b93e0d3ec6c3fc2bca811
GitOrigin-RevId: de37f83051a6ef97115564084eb3c2541edb0c58
Change-Id: I4578460cb207b37a105390410c85da6bbb125269