[roll] Roll fuchsia [starnix] Allow clone3 to specify stack.
This CL fixes a bug in starnix's clone3 implementation that didn't allow
it to use a different stack.

In the regular `clone` system call, the user has to specify the topmost
byte of the stack. In `clone3`, the user has to specify the *lowest*
byte of the stack, as well as the stack size. There was a bug in the
clone3 implementation, that it didn't take the size into account.

The new implementation of `do_clone` now sets the stack to the provided
address + the size (note that in the clone case, size will be 0).

Co-authored-by: mvanotti@google.com
Test: fx test syscalls_test
Original-Bug: 120202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794262
Original-Revision: cfd6c8f4e81d2dc114329c020043aa3afa03b829
GitOrigin-RevId: 04dcc77c02a326195b20978ee6ec429712ba7679
Change-Id: I78d745e47fe6e2a1d87dcdc59555ad0ec3601979