[roll] Roll fuchsia [gnsdk-fidl] Add new C++ bindings to fidl_library
This adds support for generating and compiling the new c++ bindings
with the GN SDK. The names of the new targets mirror the names used
in-tree.

HLCPP bindings are included with the _hlcpp suffix, and they are also
aliased to the target without a suffix until all usages can be migrated.

Original-Bug: 115033
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793915
Original-Revision: 987c872c136ea018a067b3f28910add17c740de1
GitOrigin-RevId: 5abc3cf959a774aa8601b2c083696e9296f9049d
Change-Id: Ia0452337c2ba38bb062415e96b66d133649216cd