[roll] Roll fuchsia [debugger] Avoid type lookup infinite loop.
It's possible to express in C a loop of type forward-declarations which
will cause an infinite loop in looking up symbols.

This adds a check to the recursive type lookup so we ensure that we only
visit each symbol once.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796343
Original-Revision: 0d1c4324f9b43f42fec8d3bb4e21c761f5ccea91
GitOrigin-RevId: 29b3be78fae3b92c0a5dc5431cfd860c23a7915a
Change-Id: Ia61fab3b5a152a26b620dc0d453dcd828a4fdde2