[roll] Roll fuchsia [netemul][test] Add test for interface offline
Add an integration smoke test which ensures that in typical environments
smart displays would be in in production, an interface going offline
then online results in connectivity being restored.

The test has variants for IPv4-only, IPv6-only, and dual-stack
environments. IPv4 connectivity is provided by running a DHCPv4 server
on the gateway realm, and IPv6 connectivity is provided by faking
RA messages.

Note that it's relevant that the interface watcher API is used to
verify that addresses can be found after the interface becomes online
again, as this guards against regressions which cause clients to the
API to be mislead into thinking that connectivity hasn't been restored
when it in fact has.

Original-Fixed: 117869
Test: fx test netstack-smoke-integration-test
Multiply: netstack-smoke-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/790982
Original-Revision: ff0eaa4c470ba28ded30842891b4ef68aa683e71
GitOrigin-RevId: a388871646c0e5562d5d2c6bf780bbe5b18017ae
Change-Id: I4fa2433756cca31b94db386012a1ae8b9da0cd6a