[roll] Roll fuchsia [gnsdk] Don't create c++ bindings for zx fidl
I overlooked this in the initial migration and it caused a roller
error. This brings the gn sdk into alignment with the in-tree sdk.
https://cs.opensource.google/fuchsia/fuchsia/+/main:build/cpp/fidl_cpp.gni;l=301

Not sure why this wasn't caught in any of the lsc builds, but I'll
follow up this CL with an attempt to add this case to the in-tree
gn sdk tests.

Original-Bug: 115033
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796990
Original-Revision: 378b11553fd30fb6cba87219c44c8053f9f8a169
GitOrigin-RevId: 7ea2a72bd21c9b97ba895d7940a512489026dbfb
Change-Id: Icfa71c23b428012f7cd83f5f53d630080dc53c0f