[roll] Roll fuchsia [ffx][schema] Remove incomplete schema command
We'll want to replace this with a version that's built on the new
subtool macros and uses a typesafe source of info for the type of the
structured output.

This also sets us up for not regressing more than the machine output
mechanism already has by enforcing type-safety on the argument to the
machine type, in some cases removing invalid machine declarations where
no attempt was made to actually implement machine output.

There will be some follow-up work to make things more correct in
general.

Original-Bug: 121214
Original-Fixed: 121213

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796088
Original-Revision: d83a329be68dafebabb34037257dbec72845f58e
GitOrigin-RevId: f6d69a213e42766a7ec670e38a9c9d487ad9042b
Change-Id: I4c38d9e8f15a8001ba0288691a60c9e7a872d6c6