[roll] Roll fuchsia [assembly] Remove 'empty' from feature_set_level's seen by platform_config
Capture the FeatureSupportLevel::Empty case early in platform
configuration, and do nothing.  For other values, convert it
to a crate-local version of the enum that doesn't contain the
Empty value.

Also adds it and BuildType to the prelude, with corresponding
cleanup in each subsystem's configuration code to remove the
unneeded (or conflicting) 'use' statements.

Original-Bug: 121192
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796722
Original-Revision: 38fff5273231410dddd620449ac9fdcfde85b840
GitOrigin-RevId: 87f832279196dc35984ef4a04101a58073c3c6de
Change-Id: Ibfe17140afe4ae8ee7def087fe85a8dca466fb69