[roll] Roll fuchsia [virtualization][vmm-rs] Emulate device busses with vm-device
This CL uses the vm-device crate to manage the virtual device busses
(pio/mmio). As compared to our C++ vmm this has different performance
characteristics.

The primary difference is our C++ implementation defers all trap
tracking to the hypervisor using a port 'key' associated with each trap.
From the perspective of they hypervisor the 'key' is just an opaque
value provided by the vmm. In practice, the vmm tracks trap information
(range, handler callback) in a structure and passes the pointer to this
structure as the port key. When a trap is generated, the c++ vmm would
then cast the 64-bit trap key into a pointer in the vmm's process
address space to have constant time mapping from a trap-packet to the
per-device emulation handler.

On the other hand, the vm-device crate provides an `IoManager` that uses
a BTreeMap to do a runtime lookup for the handler for a specific address
range. This is in practice an O(lg(n)) operation with respect to the
total number of traps registered on the bus.

This CL does not include any performance optimizations around the port
key but these should not be too hard to add in the future. For example,
we could maintain an additional vector of device handles and assign port
keys based on the index in that vector. This allows us to get O(1)
lookup without having to drop down to using raw pointers.

As a first step, this CL also connects the vcpu to the device manager to
handle pio writes. Reading from the pio bus and all mmio accesses will
be implemented in future CLs.

Original-Bug: 102872
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/789068
Original-Revision: 9f97059224e2ac1ff398c2b7a46d544afc09a690
GitOrigin-RevId: f0d3b3e2711dca51bf3d41dee095a8616f7f8ece
Change-Id: I889b331e75e8e5c2b01e5de485d799482035e48d