[roll] Roll fuchsia [perfcompare] Remove non-functioning wrapper for running perfcompare_test.py
This wrapper no longer actually runs perfcompare_test.py on Infra,
because FUCHSIA_DIR isn't set when the wrapper is run.  This probably
stopped working when the builder/tester split was implemented in Infra
recipes.

The wrapper was originally added in this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/305128

It was based on this wrapper:
https://fuchsia-review.googlesource.com/c/fuchsia/+/276943

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795680
Original-Revision: 414b8742b01f7ae3a88b7b6a617af72240e127a0
GitOrigin-RevId: 01f125d686edcfc2a0ad8a47dee8039256894f58
Change-Id: I5b837f12079d005257223887d409db4824d174ef