[roll] Roll fuchsia [async_patterns] Receiver and Callback
The primary motivation is to support calling methods that return a
value on some object managed by a DispatcherBound.

When a parent object manages a child object via DispatcherBound, these
two objects are intended to live in different concurrency domains.
Meaning the child could back a response to the parent as the parent
is entering its destructor. In order to make this safe, we make the
child send responeses via task posting, and passivate the tasks if the
parent is destroyed. At the same time, we require the parent to live on
a synchronized dispatcher. This way the destruction of the parent will
be synchronized with those posted response tasks. All these is
encapsulated in an `async_patterns::Receiver` object, documented in the
receiver.h.

The restrictions on sending arguments is carved out into a dedicated
header, and used by both Receiver and DispatcherBound.

A unit test is added to DispatcherBound showing how to call methods
asynchronously with non-void returns.

Multiply: fuchsia-pkg://fuchsia.com/async_patterns_tests#meta/async_patterns_tests.cm
Original-Bug: 119641
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792456
Original-Revision: f40877cd0f23ca7f5f11d5b7a543fa0cfdd5d6c0
GitOrigin-RevId: 14176ac8765d850bf48a4cf70a841095bf86c741
Change-Id: Ifb432a0781b376819f3c8a89952a3cbb6e8940e8