[roll] Roll fuchsia [starnix] Change mounts to component_mounts
This updates the configuration values for mounts that are parsed on
behalf of a component running in a galaxy. This prevents collisions
between the galaxy and component configurations, which is especially
useful in tests, where the .cml of the test contains both galaxy
configuration and component configuration values.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797982
Original-Revision: 0087d8097cba98401f44139db95d84f1eb8487f1
GitOrigin-RevId: 5b17a803136e9687a6845347a7b3c43b1145286f
Change-Id: If8cd9e47a52bbf539e5a0772e60c56b78eb26f16