[roll] Roll fuchsia [reachability] Fix criteria for pinging internet
When there has been a neighbor discovered that is not a gateway,
and the gateway is not pingable, check to see if the interface
has any default routes prior to attempting to ping the internet.

There is no reason to attempt to ping the internet without default
routes because we don't expect it to succeed.

Add new test to cover case when we have discovered a neighbor and
have default routes but haven't discovered gateway and cannot ping
the gateway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797486
Original-Revision: 89a54d1d6ca153e7ec24aee606f8019569edb166
GitOrigin-RevId: f49a51d678e401b12db06c7295b172381fb5db22
Change-Id: I3b81ba3c3409363735d558a2dab4c68a459ca7cf