[roll] Roll fuchsia [zxio] Maintain valid LRU on cache update
This change fixes two aspects of the LRU cache eviction behavior in the
RouteCache such that it always:
 - removes out-of-date entries from the LRU list
 - only moves validated entries to the front of the list

Currently, if an entry in the RouteCache is invalidated and then a
client attempts to retrieve it, the refreshed entry will overwrite the
invalid one in the cache, but the invalid entry will not be removed from
the LRU list; effectively a duplicate will be added leaving a dangling
node. This can eventually lead to a panic when that node becomes the
least-recently-used and is removed, and the cache tries to also remove
the (nonexistent) corresponding cache entry.

To fix this, always remove a refreshed entry from the LRU list before
overwriting it in the cache.

Also, only mark an entry as the most recently used after its
eventpair(s) have been validated.

Original-Fixed: 119098
Test: zxio-unittest-package
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796130
Original-Revision: 9289e6007f2ae0014b61162f71f853e0986eb59f
GitOrigin-RevId: 68161f0ebb477f51825a3b00dd0a30291feacc09
Change-Id: I387a8d99da182a18efb468ee371bf3ad1444e274