[roll] Roll fuchsia [fake-clock] Ignored deadlines never expire
Add a new API endpoint for the fake-clock to make deadlines never
expire. If a user of fake-clock calls ignore_named_deadline prior
to adding a named deadline, the deadline will have its expiration
to infinite.

Includes tests to ensure that there is a no-op when trying to
ignore the same deadline twice and that the deadline is properly
overridden.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/792731
Original-Revision: a0947ee1f66258d4aabd89c276ae2a2ff34fa27c
GitOrigin-RevId: 57f88d110f3a3de3ecd7d3ac5100a3354789f1b2
Change-Id: Ib65492ca6b32f48b9f35e77f9ecfa28275d46a5d