[roll] Roll fuchsia [netstack] Fix TCP close deadlock
This change renames `s.linger` to `s.unblockLoopWrite` and makes sure
that `s.unblockLoopWrite` is always eventually closed when closing a
TCP socket.

`s.linger` was previously left unclosed when closing a TCP socket with
an empty zircon socket, but this results in close blocking forever. If
loopWrite completes a write to the gVisor endpoint such that the gVisor
endpoint's send buffer is filled up exactly and the zircon socket is
emptied as a consequence of the write, the following (eager) attempt to
write to the gVisor endpoint will fail with `tcpip.ErrWouldBlock` even
when the zircon socket empty. In the handler for that error, we select
on `notifyCh` (which is only signaled when space opens up in the send
buffer or the socket is terminated by the peer) or `s.linger`. Since
we previously did not close `s.linger` when closing a TCP socket with
an empty zircon socket, `notifyCh` is the only way to unblock loopWrite
which is not guaranteed to happen as it depends on the peer.

https://fuchsia-review.git.corp.google.com/c/fuchsia/+/795844/11 has
details for a (britle) repro with logs showing the deadlock between
closing and loopWrite.

Original-Bug: b/233087394
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796904
Original-Revision: b811dfcf6593caed36bd2a9db91c44bd8cabf7c2
GitOrigin-RevId: 4d4a12e531ba272b2237278f99f236e2a451407a
Change-Id: I1e65bbe0eeb1a6000fec54aa1ef5c34aa0973102