[roll] Roll fuchsia [overnet] Update circuit routes when new connections occur
In order for a circuit node to connect to another node, it needs to
receive a routing update identifying that node and how it can be
reached. In order for a circuit node to *be connected to* by another
node, it doesn't need anything; a connection comes in, identifies the
node it came from, and all is well. This can occur if a new node's
connection request to another node manages to race ahead of the routing
updates produced by that node joining the network. As a result, a node
can receive a connection from another node while not being able to
connect back to that node. This is bad for Overnet, which needs to be
able to freely establish connections to move handles around.

The solution is to add a route to the routing table when we get a
connection from a node we don't recognize. We have to score it low on
link quality because we don't know how fast it is without the quality
information that comes in a routing update, but as long as it's present
we can connect back to any node that connects to us.

Original-Fixed: 120884
Original-Bug: 119353
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797149
Original-Revision: 4d565cc90e29ccb4463df598f30b8032dcb85a72
GitOrigin-RevId: 0afd62ca45c3b3e2dc9f5282bf42ec719bf43b5a
Change-Id: I1c18b7776f1eb9429353d8868760717e37e87c25