[roll] Roll fuchsia [ffx] Add tracing::instrument attributes
Adds spans to functions that use tracing.

The primary advantage is that log message will provide the name of the function. A secondary advantage is that, when we switch to a more sophisticated log-viewer such as Jaeger,we will automatically benefit from the span structure

(Also removes "level = info" where it was specified, since that is the default.)

Original-Fixed: 119970
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797943
Original-Revision: 84751022a2a211aa4875ace5dbe79d5f366fb4f9
GitOrigin-RevId: f047219673835381c0021528c349b58ac32e93cd
Change-Id: I01f09a84808bd6a8f86d3e71acbae7ca005acfff