[roll] Roll fuchsia [ffx][debug] List symbol index using JSON.
This makes the output of ffx debug symbol-index list machine-readable,
which allows developer environment setup logic to convert it into flags
and arguments for CLI utilities that can use it.

For example, the work in fxb/111998 would use this to produce
--debug-file-directory flags for llvm utilities via jq in fx wrappers,
and it could similarly be used to configure build ID directories for
GDB and/or LLDB for kernel use.

Original-Bug: 111998
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798382
Original-Revision: c53c58609fdbf8c34c039a8117581174b8e018fa
GitOrigin-RevId: 8914526bd4bbb2e7a9d2032b5cae2052846ae286
Change-Id: Ib11a8d078f04765e42eec87562b0dec75cbcf04e