[roll] Roll fuchsia [fxfs][storage] Add caching of object handles to FUSE-Fxfs
Add an in-memory cache to store the object handle when create/open is called on a file and remove the object handle from cache when release is called on a file. Every cached object handle is stored with a counter to ensure that each release matches a create/open.

Original-Bug: 117461
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795486
Original-Revision: 9886ca1c294266a639e7396ecc219ae662df35f7
GitOrigin-RevId: 6eda1063ecd352e66028b481dd3b7891efb996f8
Change-Id: I75a1c42b676b9145273b030c453e23f642579c86