[roll] Roll fuchsia [check-licenses] Gate license analysis behind a flag.
License analysis happens in CQ.
We don't need to also check them when generating license artifacts.

This CL introduces a new flag: runAnalysis. When true, license pattern
matching and result tests are enabled for the current run.

Currently, this flag is set to true by default. This needs to be
explicitly set to true in static-checks before I can make the default
behavior "false".

This CL also gets rid of the "World" package in favor of a simpler
strategy of using the packages directly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795041
Original-Revision: 728db8ef639356181232b15c150ab855eb9cf6e2
GitOrigin-RevId: 375e89f98065aa140a9344afcf7cdf54f6efbe6b
Change-Id: I5b0064cc1a9899cdab83acc77c6a60eef488f38f