[roll] Roll fuchsia [fixit][storage] remove some smklein TODOs
Remove some TODOs attributed to smklein which we are not going to do.

They are removed because they are stale, or test coverage is fine, or
because things have been working fine without the fix, or they are
related to minfs so we are never going to fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799062
Original-Revision: 6ca6a069f00399805759fae5a607bc5c5fc78465
GitOrigin-RevId: ea579c14030c4f6f7696870ee87c5f7333102e17
Change-Id: I17dc2ca816826055730adf0da68159a360fa4ba6