[roll] Roll fuchsia [kernel][vm] Simplify traversal return code handling
The ForEveryPageAndContiguousRunInRange helper in turn calls
ForEveryPageAndGapInRange, supplying it with a per-page function and a
gap function. ForEveryPageAndGapInRange is capable of handling
ZX_ERR_STOP in these functions correctly, so we do not need to handle
it in the page and gap functions used here.

Also fixes a bug where contiguous_run_len would not get reset to 0 in
the case of early termination due to ZX_ERR_STOP, invoking
contiguous_run_func on the last range twice. The added test case covers
this.

Test: vmpl_unittest
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799047
Original-Revision: 50e6065359a71f9bd8dcf735db0e21c570bfcbb5
GitOrigin-RevId: 5300baf2b26dd92b48e7d92301f8aac01c8ffe46
Change-Id: I9bb06dc8bd4262a1ae06de551000069ec523bf61