[roll] Roll fuchsia [ffx] Delete ValueStrategy and ConfigValue type conversion for Map type
For a ConfigQuery, to get a Map<String, Value> typed config value, it
will call `validate_type::<Map<String, Value>>` on the map recursively,
which tries converting every single value (even if it's not a Object)
in the map to Map<String, Value>, and it will fail.

So every ConfigQuery::get<Map<String, Value>> call will only return
an Error instead of the desired value. The only correct way to get a
Map<String, Value>> is to get a Value and then check if it's a
Value::Object.

The code is currently not used anywhere so it's safe to delete it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799206
Original-Revision: a3e7f21eb6c614674da6f261486f0e2b0c01d12c
GitOrigin-RevId: 94ac053b0cc13f2121e1118eecdd4ce7427733e7
Change-Id: Ied9258249c3cf755cce0b15071ba4d1b0407e40a