[roll] Roll fuchsia [dhcp] Don't use std::mem::size_of_val
std::mem::size_of_val is a footgun, because it is easy to accidentally
use it to measure the size of a Vec (and get back 24 regardless of the
Vec's length), when you meant to measure the size of the slice contained
by the Vec (which is the number of bytes taken up by the data stored in
the slice).

This was causing us to incorrectly serialize the ParameterRequestList
and PathMTUPlateauTable DHCP options.

Test: dhcp_protocol
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799684
Original-Revision: c3f5d58f6f6b653ae580c70b10d931e30bc71c8d
GitOrigin-RevId: d9cd9accc2eb092eb749743a9f0c1a3671b0427b
Change-Id: I953257a34bd63dd726e9354192e317b58443c64d