[roll] Roll fuchsia [cm] Update RealmQuery protocol
This change is similar to the one done for
LifecycleController in https://fxrev.dev/787842

We want to remove the existing methods from RealmQuery
and RealmExplorer and replace them with better methods that
support new features and precise error types.

To keep with platform compatibility guidelines, we are making
the following changes:
1. Add new methods and mark as available in API 11.
2. Mark existing methods as deprecated in API 11 and removed
   in API 12.
3. Add `legacy = true` annotations to existing methods.
4. For API 14, we will remove the `legacy = true` annotation
   from the existing methods and remove their implementation
   in component manager.

The result is that `ffx` binaries targeting API 11 or
earlier can use the existing methods on platforms supporting
API 13 or earlier.

All clients of RealmQuery and RealmExplorer targeting API level
12 or above are required to use the new methods.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797883
Original-Revision: 30e52446f7d8581a29e023e3fa642a2333b8e623
GitOrigin-RevId: 8225bb05dee9e64c6c8123c557583fdb208847a9
Change-Id: I886502006e2cdc58ea0e09eda2b60902ab2fc1e3