[roll] Roll fuchsia [starnix] Fix binder driver refcount issue.
The binder driver allowed for weak and strong refcount operation to be
re-ordered. It means that the following operations:
- IncStrong IncWeak decStrong decWeak
could be reordered into:
- IncStrong decStrong IncWeak decWeak
This is invalid, as the remote object would be freed after the decStrong
command.

To ensure this doesn't happen, the binder driver will not send any
decrement command until all increment commands (weak and strong) are
acknowledged. This ensires in the previous sequence that the decStrong
command cannot be executed beffore the IncWeak happened.

With this, the binderLibTest is not crashing anymore on an use after
free.

Original-Fixed: 120058
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800482
Original-Revision: 9ade6a27b44e645c5fc085650cafcdeb4317dbd5
GitOrigin-RevId: 0ed5c5b53fc6918bf0b8e0e8ea28d59573779f25
Change-Id: I60d5cbb14895c14130be7d29b7058050d03989eb