[roll] Roll infra/recipes [gerrit_auto_submit] Don't change owner when rebasing
Gerrit now supports rebasing "on behalf of" the change's current
uploader, which should let us fix the longstanding pain point resulting
from a rebase by auto-submit taking away the change author's implicit
OWNERS approval of some files affected by the change.

The easiest way to test this is to just roll it out and watch closely
for breakages, since it's not the end of the world if auto-submit is
broken for a short amount of time.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/800583
Original-Revision: a3f4252c1e8e37915c2d8ca8926487c5c585b99b
GitOrigin-RevId: aa85a0ce26384b54fdce0b85ebfd6dd75c0a2f04
Change-Id: Iff4c3e70624caf660fa42196b3aa869d6ac26512