[roll] Roll fuchsia [base-resolver] Swap in the true-base-resolver
Replaces the current base-resolver implementation with that of the
true-base-resolver (which serves package directories directly from
blobfs instead of relying on pkg-cache's pkgfs/packages directory).

Uses this new base-resolver to resolve pkg-cache and so deletes
pkg-cache-resolver.

Serves the shell-commands "bin" directory from base_resolver instead
of pkg-cache. console-launcher doesn't add the "bin" directory to the
shell's namespace until the "bin" directory starts responding to
requests (to prevent hanging the shell with unresponsive namespace
entries). base_resolver using pkgfs/packages caused pkg-cache to start
early in the boot sequence, but with this change there are some
configurations where the shell accessing the "bin" directory is the
first use of a pkg-cache capability and so is when CF starts
pkg-cache. On asan builds this is slow enough that some tests fail
because "bin" isn't added to the shell's namespace soon enough and so
some shell binaries cannot be found.

Original-Fixed: 101492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794970
Original-Revision: 4399948e198a00d746cd65e0645df7150b1982b8
GitOrigin-RevId: ed9e497ae746f254f2ec6fca4c21f5c5aadcef13
Change-Id: Id2ff33c7de3bf0a11766bcbe1ef26aab585d2dff