[roll] Roll fuchsia [netstack] Factor out common exposes from shards
Netstack3 and Netstack2 are both used as the system netstack. That means
they need to have a common subset of exposed protocols. Factor that list
out into a single file so that it can be shared between the two, and so
updates can go in one place instead of two.

Original-Fixed: 121408
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798044
Original-Revision: 780fbeeaef431ab055706f076a357bacc0c7a844
GitOrigin-RevId: f6a097076f328cf468e2828912dfde02a2762283
Change-Id: I6c758aec52c1f779ab19ecb704b8710d6733816f