[roll] Roll fuchsia [zircon][docs][fixit]Fixing markdown syntax in boot-options.md
This is a reland of fxr/799643: Revert "[docs][fixit]Escaping brackets which are not ref links"

This escapes the [ so that they are not interpreted as auto reference
links.

I also added comments to the generated markdown to point to where
the file is generated.

Original-Bug: 106059
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799822
Original-Revision: e7159e03bde4e390783a4219fc2129486dbd765a
GitOrigin-RevId: 40ba276e544ad00da3490d7c2d52b7fad7ce9cf3
Change-Id: Ib548e49d91759b8ef57a163d3a9cb30effe9b3b1