[roll] Roll llvm-project/libc [libc][NFC] reorganize structs in printf
Previously the type description structs were defined in the parser. For
the fuzzing targets we'll need to use those, so I've moved them into
core_structs.h. Additionally I've renamed the function for determining
the TypeDesc from a given type. Previously it shared its name with
get_type_desc which is a related but separate function that is a part of
the parser.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D143595

GitOrigin-RevId: 3565aaae328de17c0b54789401e4b2a92f4cd4e3
Original-Revision: 99afcb3c902b755ad9f368e3131f472802f67e86
Change-Id: I0aa33a8486da0866d5ae132b413d1f9e5ba0d483