[roll] Roll fuchsia [bazel_sdk] Replace debug-symbol related bash scripts with python ones.
This replaces elf_strip.sh and generate_symbols_dir.sh with
equivalent Python script. This is more portable and removes the
last use of the `file` command which is problematic on CQ bots
and remote builds.

This also introduces an `elf_info` Python library to query ELF files
(only ELF64 is supported for now), used by both scripts.

This does not seem to slow down local or remote builds, but fixes
remote builds which failed because of the missing `file` command.

```

$ fx bazel clean && \
  fx bazel build //build/bazel/examples/hello_cpp:pkg.debug_symbols

  BEFORE (Best of 3 runs)

    INFO: Elapsed time: 17.964s, Critical Path: 11.16s
    INFO: 457 processes: 305 internal, 152 linux-sandbox.

  AFTER: (Best of 3 runs)

    INFO: Elapsed time: 17.587s, Critical Path: 12.93s
    INFO: 457 processes: 305 internal, 152 linux-sandbox.

$ fx bazel clean && \
  fx bazel build --config=remote //build/bazel/examples/hello_cpp:pkg.debug_symbols

  BEFORE (Best of 3 runs after warmup one)

    INFO: Elapsed time: 11.237s, Critical Path: 9.52s
    INFO: 457 processes: 152 remote cache hit, 305 internal.

  AFTER (Best of 3 runs after warmup one)

    INFO: Elapsed time: 11.626s, Critical Path: 9.99s
    INFO: 457 processes: 152 remote cache hit, 305 internal.
```

Original-Fixed: 99458
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/795084
Original-Revision: e85c796c0faa2643725d966486a383387594aef1
GitOrigin-RevId: 3e23a688993dd540fa80dc2c084ac8dc876c39e1
Change-Id: I7e86c96b9fe2ae6b6a790de302fb0af96441bdfa