[roll] Roll fuchsia [fxfs] Enable LTO
Re-land of fxrev.dev/791112 with a slightly different approach, which is
used by ffx and should work.

This gains back the performance which was lost when splitting Fxfs into
several crates.  See, for example, WalkDirectoryTreeWarm.

The cost is a ~45% slowdown in compilation time for fxfs (19s -> 33s on
my machine).  If this ends up being problematic for productivity, we
could further split up Fxfs into a binary variant used for tests (which
won't have LTO) and another used for benchmarks and production (which
would have LTO), but that doesn't seem necessary given the build time is
still reasonable.

Note that LTO is not applied to the unit-test suite, so building the
unit-test suite won't be made slower.

Original-Bug: 119533
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799622
Original-Revision: 2127d5374e60c9f9d064fb567d3759116de1c4dc
GitOrigin-RevId: 982d2bf71fa44c8f5688520e2e9283d38d831522
Change-Id: Id4d1588d55733d2383150d89a19862cc50ffa9b6