[roll] Roll fuchsia [wlan][policy] don't retry short connections
When a disconnect is detected from SME, the state machine tries once to
connect to the same AP. After this change, the immediate connect
attempt will not happen if the connection uptime was short. Instead
it will fall back to network selection where another AP may be chosen,
and the previous AP might be avoided because of the short uptime.

The reason for this change is that we have seen some AP's repeatedly
disconnect us, and we keep successfully connecting to the AP. We should
go to network selection and choose another AP, which may not do the same
thing.

Test: added unit test
Original-Bug: 120610
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800863
Original-Revision: a25e22460a87346db36da4b62ac701e5033911f3
GitOrigin-RevId: 37a5048e639d91f2963f6b7a37bab24956804a53
Change-Id: I974d1e4b0ad0877e1e15f523c81bebb8932b0ba5