[roll] Roll fuchsia [starnix] Escape file names in /proc/../maps
Newlines in the names of mapped files should be replaced by "\012" in
maps output so that the contents can be parsed unambiguously.

This change is mostly refactoring tests to permit running tests for
/proc behavior both on Starnix where /proc is not always mounted and on
the host where /proc is mounted and the tests are usually run without
root. This introduces a ProcTest fixture that will set up a proc
filesystem somewhere if needed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801709
Original-Revision: 5082a88aaafbcfb6c93195ff682d778027a371be
GitOrigin-RevId: 40ab1c464b53c82af0598f093153b77cf6e0b531
Change-Id: Ibaf4c37cb51ea8154f546a433bfaca35a6e8d5be