[roll] Roll fuchsia [tefmocheck] Only check ffx timeout error on failed tasks.
If ffx fails to connect to the target, we retry and if it succeeds, then
we keep testing. Tefmocheck should ignore the timeout log in these
cases. If ffx continues to fail to connect to the target, it should
return a fatal error and fail the task, in which case tefmocheck can
assign it this failure mode.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802008
Original-Revision: d28b13bafc0f4fda6f688abdce97e3b039e7111b
GitOrigin-RevId: 1281bf4e4700d72b2983868e849e1ebef9a66ef6
Change-Id: Ib4c27206f3c7c0bb178e9d4bf7642f8aa462ae45