[roll] Roll fuchsia [ffx][writer] New writer library
This implements a new type-safe writer library for FHO-based subtools,
that more or less resembles the old interface but enforces a single type
as its output.

This will allow us to produce much more useful schema information when
we get to it, and ensure that in machine mode tools are only outputting
the type of data they claim to support.

Original-Bug: 114369

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797523
Original-Revision: 0d09cdb1dc4ac97c27863ca2a0808532e00cac85
GitOrigin-RevId: 482f0edf471e29034bbafb8ccaf8f6a093759f31
Change-Id: I56687ab6eaed530b525f2e329d71034fcec6250e