[roll] Roll fuchsia [check-licenses] Don't run license analysis by default.
In a previous CL, a flag was added to check-licenses that
allows the tool to skip over analyzing license files and
running tests at the end of the execution, saving several
minutes overall.

This CL sets that flag in the license_data.gni template, so
users of this target run significantly faster.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799982
Original-Revision: ff2b0f14464a748edb63d6e8cc83a29054af9187
GitOrigin-RevId: 23e44e5fe84cc292b9825bbc6a57fbdc9fcce7e5
Change-Id: Iac21b1ade270a56c265b89d7fecbf47353fd9009