[roll] Roll fuchsia [libc] Add macro for disabling SCS and hwasan instrumentation
The core-c11-thread test would fail with
https://reviews.llvm.org/D139377 because that patch would cause the
__hwasan_load/store functions to be non-leaf functions, meaning they are
instrumented with SCS. This particular test would fail with a page fault
because we could end up in the state where a thread is explicitly
detached after it has exited, meaning any hwasan checks done in the
detach call will call into hwasan and access the SCS even thought it was
deallocated when the thread exited.

This isn't indicative of anything funadmentally wrong with hwasan, but
more happenstance that we use runtime calls for access checking with
only hwasan. It's also only in the particular unique situation when we
can call into an SCS instrumented function after SCS was deallocated.

This CL adds a macro for explicitly disabling SCS and hwasan on specific
libc functions where this case is seen. It also replaces all libc uses
of __NO_SAFESTACK.

Original-Bug: 115924, 57112
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/766781
Original-Revision: 58986a79f3173fa3bb383f336df3828ca1f3d22d
GitOrigin-RevId: 9bf9e3467eb9a0d73f755c7c785e9272f9efcbfd
Change-Id: I748645b0982bfcded9ed5d5461280504925933bc