[roll] Roll fuchsia [sysmem2] change display_size to display_rect
Since some codecs support cropping on the top left, it makes
sense to use a rect with upper-left values as well, instead of
a size with only the width and height.

Nothing about this CL should be taken to imply that we intend
to support top-left cropping in any pipeline that doesn't
actually use it. While non-zero top-left is possible, it's
definitely not common. Ignoring these fields when they're not
actually zero would look wrong, but wouldn't be fatal, and
would be no worse than not having the x, y top-left fields.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796763
Original-Revision: 76ba2fd2d4705b9fbb7619468d6e1fdf37621efb
GitOrigin-RevId: c934a7c8f94c0714a5669d757e6fe6638816daf8
Change-Id: Ic70c8615b10b1471edd72a6b627f839d2dfd76f8