[roll] Roll fuchsia [fxfs] Fix lock race on rename.
There was a race where if there was no target file at first check it
would remove the object from the directory without having the correct
lock on the target that arrived there after. Now the acquire lock method
now will always return the transaction, but indicates if the object is
there or not while holding a lock on the directory to prevent that from
changing.

Original-Bug: 119138
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802428
Original-Revision: cb1ba9e501f24264048911a4f43a70e3c3b13c9d
GitOrigin-RevId: 3f02e56900e81a268fdd56845156d93812eb48ad
Change-Id: Id69dc90417eac875815e8f5deda91fd3325062d0