[roll] Roll fuchsia [netsvc] Avoid opening /dev with rights
Dealing in channels has the effect of eagerly encountering errors, which
required some restructuring. In particular, the previous code opened
"/dev" as a file descriptor which works even when the namespace contains
"/dev/class/network" such that "/dev" is a local node in fdio; this
would later fail in `Paver::OpenDataSink` which tries to
`component::ConnectAt` through this file descriptor which might not be
backed by a channel - it must have been the case that this code path is
not exercised when netsvc is running in core (as opposed to bringup).

Original-Bug: 101092
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802072
Original-Revision: ff0ebfee7261140bdf3c1595ee4e2a7284d2f214
GitOrigin-RevId: 0bb58e87694d2f280449dfc7eab3e75f9860cdb0
Change-Id: I44a8ceff1bb433086b5c3a1daffb71b116055b7f