[roll] Roll fuchsia [ffx][overnet]: Cleanup ssh connections
When dropping an Overnet host pipe connection, explicitly kill the
spawned ssh connection. Previously this was not killed when the target
was dropped from the colleciton, causing zombie ssh connections to the
target to persist. This would require a daemon kill to remove (or a kill
on the ssh processes themselves)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802188
Original-Revision: a1b4ee1d9ecbb512a8ba5d4923d146f25369dbc6
GitOrigin-RevId: 3154d9c27b9ddf87a0f3d40c0ca709db0e9dcc39
Change-Id: Ifbb37181ede45e8a94b62f9ce4c9c341df44460c