[roll] Roll fuchsia [display] Close BufferCollectionToken if import is not supported.
This change changes the default behavior of banjo ImportBufferCollection
method -- if it is not supported, the driver should try to close the
BufferCollectionToken. Otherwise it will trigger a ZX_ERR_PEER_CLOSED
in sysmem and cause a failure on all the other shared tokens.

Original-Bug: 121411
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803770
Original-Revision: 06fbd5a6bddb1f95e414c277eb1795c6c7e50686
GitOrigin-RevId: a1ce3674ee5a60e06dbc44d3d80328bf46b0b1b9
Change-Id: I805e0b4745b94556de4b8b6149412e53e2da272c