[roll] Roll fuchsia [amlogic-display] Implement ImportBufferCollection / ReleaseBufferCollection.
This implements the Import/Release BufferCollection banjo API to
allow clients to import a BufferCollection token.

The driver will need to maintain a sysmem Allocator client so that
it can use the Allocator to bind the token to a BufferCollection
client.

Test: Added fake sysmem test fixtures to amlogic-display-test.
Original-Bug: 121411
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802309
Original-Revision: 9d1e895f696bf99d8725fbec6786860d9274e754
GitOrigin-RevId: ddfde8e94e87a7bdfb4ec47eeda0249ba7f2ee0d
Change-Id: If52521066109c3d211cc65a556a0de433c1abf9e