[roll] Roll fuchsia [intel-i915] Implement Import/Release BufferCollection.
This implements the Import/Release BufferCollection banjo API to
allow clients to import a BufferCollection token.

The driver will need to maintain a sysmem Allocator client so that
it can use the Allocator to bind the token to a BufferCollection
client.

Test: Added fake sysmem test fixtures to intel-i915-test.
Original-Bug: 121411
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802312
Original-Revision: c1a1135c6ed23d94e4cf46758a4242461a719d35
GitOrigin-RevId: 9de93ff77aec5f4148f085c4fc6f1872e24fba64
Change-Id: Id0426e9a9ba3348bb537b65960a939ebe4b671a9