[roll] Roll fuchsia [ramdevice_client] Replace get_path with as_dir
Rather than exposing the block_path of the ramdisk, we now force
callers to interact with the ramdisk via either a DirectoryProxy or a
ControllerProxy, only providing the DirectoryProxy when necessary.
There is further work to be done here with regards to multiplexing
between the Controller and Block protocol.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/794202
Original-Revision: 2d4bbd1f3766dd4a3171b509825f3de4749eb451
GitOrigin-RevId: b373bcb49151cf0591dff0a91ab2f53a1660779d
Change-Id: I0c5c769554368c610d333dc1f210ce103d283059