[roll] Roll fuchsia [starnix] Fix ffx starnix adb command
Previously the kernel was not serving the galaxy controller in the
outgoing directory of the galaxy. Instead, it was only serving it in the
outgoing directory of the kernel itself. This used to work, when the
galaxy and starnix kernel were the same component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804127
Original-Revision: 4e0f1acea7b6c20a7c874d1e2ca7e5f87c94c953
GitOrigin-RevId: a2dc4ebb325b11f85291854c0d12454a6fe08af2
Change-Id: I0ade6abd2b55e9d099dd73616fe925d3de5f4ce6