[roll] Roll fuchsia [bt][avrcp] Fix packet decode issues for media player feature bitmask.
The issue was occurring due to incorrect endian-ness interpretation
between the rust bits and decoded feature bitmask number.
AVRCP packets transfer multiple-octet fields in big endian. The lowest
octet/storage address feature flags are therefore the big end of the
feature number that's transferred. The FIDL bits is modified so that the
feature bitmask can be correctly interpreted into the feature flags. We
make the changes on the FIDL side and not the packet decoding side,
since it's
easier to compare the values against the spec if we keep the decoding to
match the endianness that's specified in the spec.

Test: fx test -v -o bt-avrcp-tests
Original-Bug: 108372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/797862
Original-Revision: 5f6bd757eb5f2dd4a486a9391f394d04a9bc909d
GitOrigin-RevId: 49646023b9f67cf85740b35f8f343561b5aa284b
Change-Id: Ic670c9e6f68d8895c89a8526c68e273d32741b27