[roll] Roll fuchsia [ffx][fho] Implement `fake_proxy` for fho
In the old `ffx_plugin` macro, it would generate a bunch of test
functions to generate fake proxies. In order to port some of the
existing plugins over, we still need this, but it can just be a generic
function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803808
Original-Revision: bf8b3f91e3ec1b53ed705e96c39e1182c07721b0
GitOrigin-RevId: c7d6bf762c52132ca4e09a6c4e16ca44903ad8a3
Change-Id: Icf2055519a7d4c5b956aeb6d77bb0b1820011c49