[roll] Roll fuchsia [netstack3] Split capacity and target_capacity
Use a separate method for reporting the target capacity of a socket's
send buffer. This fixes a bug where, if a TCP socket was full, a
reduction in its send buffer size wasn't reflected by the next call to
read its send buffer size. Add an integration regression test for the
behavior.

Since the test makes additional use of the ability to set and retrieve
the send buffer size for a TCP socket, add methods for such to the
fuchsia_async crate.

Original-Fixed: 121449
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802347
Original-Revision: 9c3eacdb417e09358604d423c84e5836da14418a
GitOrigin-RevId: c7e54d6bdd82cc3166430d8166f25e179956e3d0
Change-Id: I7ac9ee33a7de0e4250cc5173be14c71be326b6fb