[roll] Roll fuchsia [seriallistener] Print logs periodically while reading from socket.
Some zbi tests seem to be just barely going over the timeout. However,
it's possible that seriallistener is just slow or hanging when trying to
read from the serial socket. Printing logs periodically can help give
timestamp estimates for when seriallistener is actually reading the logs
to determine if the bug is in the test itself being slow or the reader.

Original-Bug: 116559, 119856
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804151
Original-Revision: 55db89326ba3ecdf9ae4f3194ae6a7a3891eb6c2
GitOrigin-RevId: 3c8d6a24d813973bebccdee52ccffe6f71d146e2
Change-Id: Ifeaebd3403010af3cb0bf5bda733b7471e363d86