[roll] Roll fuchsia [session-manager] Add a core shard for `minimal`.
A variety of protocols listed in 'required_capabilites` aren't
necessarily available on products based on minimal.gni. So I've moved
them to "optional_capabilities" and tagged them with
`source_availability: "unknown"`.

I also renamed `required_capabilities` to `all_products_capabilities`
and `optional_capabilities` to `some_products_capabilities`. The use
of the word "optional" in the name of that file contributed to my
incorrect assumption that the capabilities listed in that file were
received as "optional" by the session.  This is maybe more verbose,
but I think truer to the intent of the files.

Original-Bug: 121597

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800704
Original-Revision: 3f1885691775a66474afde90756c792396741d57
GitOrigin-RevId: 5dacbbb14f50c1a4aaa3fdb188935aa1bc800c97
Change-Id: I065f701110a740a88f072be1d0b3d21da888c874