[roll] Roll fuchsia [botanist][ffxutil] Set config values before starting the daemon.
The daemon might not be picking up new config settings if they are set
after the daemon is started, so we should try to set it before where
possible.

I also added a function to get a copy of an FFXInstance with a different
associated target so that we don't need to duplicate the number of
`config set` calls for each target.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804331
Original-Revision: 94c73dd3e2889dc78bcff9b67d24e208c4bae01f
GitOrigin-RevId: a34ab11d2daeb9e909f71589c69ee58cc507e28b
Change-Id: I419bdb3194e0278fe58654fe0593d58d04b28c48