[roll] Roll fuchsia [sl4f] Dart null safety: convert more tests
The changes to test files here were all produced by "dart migrate".

These test cases all pass a null 'ssh' argument to the Sl4f
constructor, so change the constructor to allow a null argument in
order to accommodate that.

Note that we're not changing Sl4f's 'ssh' attribute to be nullable
because we don't want to have to change the various use sites to use
".ssh!" to access it.

Original-Bug: 120221
Test: fx set terminal.x64 --release --with sdk/testing/sl4f/client:tests
  + fx test sl4f_client_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802970
Original-Revision: b64f83d3ec2b2fb9116e3b4e410674072c45751d
GitOrigin-RevId: 869cccea5e099b1ccd4ac9374ed7b23ad928566b
Change-Id: I6554652e8d85b367a97a61c2c0de7a78215e58ed