[roll] Roll fuchsia [dlog] Remove dlog_force_panic.
There are two separate and orthogonal ways of representing a panic
within dlog. The first one consists of simply bypassing it, by
overriding the boot option. The overall effect is that printfs
will not attempt to write to dlog if we are bypassing it.

* dlog_force_panic(void).

The second form of panic, consists on a flag that is actually set in the
panic path. This flags causes writes to dlog to bail out early with
'ZX_ERR_BAD_STATE'. In this case, printf will not bypass dlog, but will
attempt to write it, unsuccessful of course. But because it failed,
it will continue and attempt to write directly to the console and
serial.

* dlog_panic_start(void).

The only user of dlog_force_panic is k wtd force, which relies
on this method for writes to stdout to go straight to the uart.

This is still the case, and is closer to what panic actually
does.

Original-Bug: 89182
Test: k wtd force
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/799026
Original-Revision: 2126d3416634da71504da11ffc33ddb70525fb2b
GitOrigin-RevId: 18f75ed35c875d86cc2edf59bd0f7972473a3960
Change-Id: Id4654da12cc141c2b53325f3844a395bfd2dfdab