[roll] Roll fuchsia [ffx][config] proxy timeout getter for Environment.
This is a manual function to get the proxy timeout from the config
environment. This is a placeholder for what may later be a more general
config value loader that doesn't rely on a global context.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804590
Original-Revision: de26f93d74cda86d7a4d3b6df145fd2d9d4341b6
GitOrigin-RevId: 494a9860e0be612985d2e979ed0b8bee0eb8462d
Change-Id: Ia728fe3e7a9a12a91ce9985c0939ddc993478013