[roll] Roll fuchsia [bazel_sdk] Remove fuchsia_only_target()
The fuchsia_only_target() function is not documented and seems to be
used internally to return a value for `target_compatible_with` argument
in rule invocations.

However, using a value of `[ "@platforms//os:fuchsia" ]` is both simpler
and equivalent to express that a target can only be built in Fuchsia
compatible build configurations.

Moreover, the fuchsia_only_target() implementation could never trigger
the error message of fuchsia_select() since it called it with a
"//conditions:default" condition label, which would always be matched
anyway, so there was no benefit in using it.

The function is no longer used internally by @rules_fuchsia, and is
not being used by downstream clients, so can be removed safely.

Original-Bug: None

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804948
Original-Revision: 94d93df37e2d06c319db4d76d1ac31fbe26a3568
GitOrigin-RevId: 9ad3c68e2e6dd1e69a29b4977e041e511a4bf2ec
Change-Id: Idfa14ad13e4d3903f8dd1a08d02ebdd3cd6eca33