[roll] Roll fuchsia Revert "Revert "[realm_builder] garbage collect manifests and local components""
This reverts commit 9f49e810e37f37dd056b6bca8f60dcd057493b44.

The original commit was reverted due to causing flakes in tests that did
not keep the `ComponentRunner` channel provided in the `Builder.Build`
call alive for the duration of the realm's lifetime.

There have been no significant changes to the approach taken in the
initial commit. The only way to identify which tests do not keep this
channel alive long enough is to land this and fix-forward any tests that
begin flaking due to this.

Original-Fixed: 121196
Original-Fixed: 121808
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802709
Original-Revision: 8d531f6dbcb7e208b6baae6a918e817e524d6680
GitOrigin-RevId: da9a3313bdc2f8b2597c63c842b1d2db7998d56c
Change-Id: I641b4557a66ab10cd4f6bb98bd96b88d24deb856