[roll] Roll fuchsia [timekeeper] Adds a loop pause for suspendable sources
Adds a pause in the sampling loop when the time source can be suspended,
as in the case of pull sources. The old loop was adequate for push
sources since it would block until the push source sends a report.

In case of pull sources, having no pause here means that we will spin
and pull time information constantly from time sources. This is
exactly not what we want with pull source. Instead, we want to limit
the number of times that timekeeper calls into the time source to pull
time from, and minimize the length of time that the time source needs
to run.

The pause was added at the top level (instead of e.g. at the level of
the individual source) to ensure that all work related to time sampling
is completed before the clock maintenance loop goes to sleep.

Original-Bug: 121920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804656
Original-Revision: a0cae2dc2ea031de297488e3d3b7f6abae8a0d46
GitOrigin-RevId: 8fea8f27aaafbd0ab940a9bc881dd86b3cc4def4
Change-Id: Id839de62ad8c5e1a777c75b068db615bc9995a8e