[roll] Roll fuchsia [starnix] Add ARM64 stat_t and trampoline file.
Define stat_t for ARM64 which is slightly different than x64. Change the
definition of the dev_t type in uapi.rs to be u64. Previously it was
__kernel_old_dev_t.

The generated uapi bindings don't capture the dev_t definition so I
assume __kernel_old_dev_t was used instead. But the structure is defined
using dev_t and not the "old" one. __kernel_old_dev_t is 64 bits on x64
so it matches dev_t, but it ends up being 32 bits on ARM64 which is
incorrect.

Adds a stub file for the restricted_enter() and restricted_return()
functions so it can link.

These functions are left unimplemented so that they can be tested at the
same time they are written. This patch adds detailed implementation
notes.

Original-Bug: 121659

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801233
Original-Revision: 1e00ecf203c1f459bd01b536cbd38389cc2e8ed6
GitOrigin-RevId: ff07f95eef593f5e25ec8caf86cd650efdb89896
Change-Id: I554a6de5bfbaa64becc346c418f939ca5d47afca