[roll] Roll fuchsia [stanix] Explicitly drop Task on exception
Previously the task reference that was in the syscall dispatch loop
would not be decremented if the thread was killed in the exception
handling path.

This is a temporary hack that needs to be in place until restricted
mode exposes exceptions to normal mode.

This change explicitly decrements the Task reference count when the task
is killed due to an unhandled exception.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806347
Original-Revision: c38b00a3fb81b99b39753b942801245f125d0785
GitOrigin-RevId: 5a8355729eb83b655342cd6363b2e2b0b90ab076
Change-Id: Ib110a607df7cbe69dd8d62c67f6c44db28ffcefb