[roll] Roll fuchsia [assembly] Remove some minfs references from assembly.
This CL originally aimed to remove minfs assembly support and replace it with "--empty-data". It has been shrinking over the past few months as pieces of it got implemented elsewhere. This is the last of the original changes, renaming Minfs to Data in assembly code and fixing usage text.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743202
Original-Revision: a8bdd6754909402939e124dfefa58e954e929395
GitOrigin-RevId: d00d4d6b0fae55c8ab962d4906d8cdca7e34810c
Change-Id: I363210df6bc934b7f944dec4f5cbd25d0377d9c4