[roll] Roll fuchsia [phys][gcc] Add PANICs to enum class switch-LUTs.
In a few different places, there were functions/lambdas which
basically acted as LUTs for enum class values implemented using switch
statements.  Even though all of the define values for the enum-classes
were covered by case statements, the compiler can still complain that
a function may not return a value depending on the compiler options
which were chosen.

Given the potential for uninitialized variable and/or memory
corruption, a demand from the compiler to define behavior of
"impossible" class enum value does not seem unreasonable.

So, in the places that the compiler was complaining, add calls to
ZX_PANIC in the case an unrecognized value is found. PANIC is flagged
as no_return, so this has the effect of suppressing the compiler
warning that "all paths must return a value".

Original-Fixed: 122177
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806228
Original-Revision: fd8805c196b4f3d03e2ad96157e311b914b3ff71
GitOrigin-RevId: 6955a9bea301cf9134443ef2cf2c358df5efaefa
Change-Id: I8b6ab0263031581d0068cc1e8382e712772bb36a