[roll] Roll fuchsia [component_debug] Manifest is encoded and sent using an iterator
The RealmQuery.GetManifest call fails for manifests that are too
large to fit in a Zircon channel.

To work around this problem, component manager will now encode
the manifest into a persistable format (per RFC-0120) and send
that data over a dedicated iterator.

We cannot use a VMO because we need to maintain compatibility
with overnet.

This change also cleans up GetAllInstances by tying the iterator
to a TaskScope and closing it when there is no more data to be sent.

Original-Fixed: 122209
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/806987
Original-Revision: cc8eabe6f9821c375595f289f8873a6b0b3e1e5d
GitOrigin-RevId: f1adbd268aa702fd7652e39ea821a6ca87443fa5
Change-Id: Ib655cfbb15d93540a409a421663d7ac0e5524890