[roll] Roll fuchsia [scenic][test] Fail louder
The test fixture ignored errors in a FIDL call to scene manager,
which results in a confusing failure mode in tests.

This makes the failure a bit louder, in the hope it will be
recognized for what it is in the future.

Original-Bug: 122116
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807449
Original-Revision: f1a009bc314e4c44c4dce2db514c312cf501d003
GitOrigin-RevId: d6f1276ad4c45ef6c1345218dfdb820c37b7825e
Change-Id: Ibaf6371f8b9182e7d2ccdeec62d46904b3ba35aa