[roll] Roll fuchsia [ffx][emu] Refactor instance API to ffx/lib
This refactors the emulator instance data methods and structs to
a library in ffx/lib/emulator_instance. This is in preparation of
updating target_collection to get the emulator instances from the source
data vs. trying to cache them in the config.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803987
Original-Revision: 45475a6996fdf47601063afa5ef70cb3c028dd55
GitOrigin-RevId: 7fc1f096ba1674d114ce11dce4be86c4dcb70a23
Change-Id: Ic8932fb6ad2b04d0a1f4734840390ceb7b70e583