[roll] Roll fuchsia [media/ffmpeg] fix broken profile dependencies
During the transition to product-specific profiles, we want the
'default' and 'max' profiles to be the same. 'max' is always used,
and we want to use 'default' instead when a product doesn't
specify a profile.

The BUILD.gn that defines the default and max profiles was using
deps in a way that was not effective (max ended up empty). This CL
makes both default and max depend on a common group target that
has the right deps.

Test: build change pertaining to ffmpeg roll only

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807607
Original-Revision: 8042a8cda7fe58bd18e9447e436a45caf7caed39
GitOrigin-RevId: 6ce135896269cf5abf74f20991651b313e2c493b
Change-Id: Idb0e67b6d500e24232d63649c9dc18d54e7710b1