[roll] Roll fuchsia [netstack2] handles route interrupts in order
Update the `routeWatcherEventLoop` to select on interrupts in a
deterministic order.

The original implementation would evaluate cases in a psedorandom
fashion (as is the behavior for Golang `select` statements). However,
this leads to pseudorandom behavior observed by clients of the watcher.
Clients may observe routes as either `Added`/`Existing` depending on the
order of case evaluation.

This CL uses a single channel to hold all interrupt types, which ensures
that the interrupts are handled in order.

Multiply: netstack-routes-fidl-test

Original-Fixed: 121918
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/804929
Original-Revision: e4109620c7874fe8de44488939e73237c6b28000
GitOrigin-RevId: ff7bb621c01609f6db43c251cbe697f9f877eb82
Change-Id: Iae0717b97216ed622589780796a366185d0c2b08