[roll] Roll fuchsia [realm_builder] Decouple teardown from destructor
The current code is confused and results in races between
`RealmRoot::~RealmRoot` and dispatcher shutdown, resulting in:

- resource leaks (dispatcher shut down, `DestroyChild` not called).
- dropped callbacks (dispatcher shut down, callback not called).

Avoid this mess by separating the synchronous and asynchronous shutdown
paths; synchronous shutdown via the destructor calls `DestroyChild`
synchronously, and asynchronous shutdown via `Teardown` takes a callback
and asserts that the callback is called (i.e. the dispatcher is not
shutdown before calling it).

Tests that use `AddLocalChild` may now deadlock if they neglect to call
`Teardown` and wait for the callback to complete.

This is an API-breaking change.

Original-Fixed: 121808
Original-Bug: 122045
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802829
Original-Revision: 4e74b6a6e4b95a175233c15ea44dfec20fab08b4
GitOrigin-RevId: 5b027c4952b43bca8bc6599ef564749056fe3a2a
Change-Id: I5afc7ed6a98eb1aefd9920db4dfdec3e7e98fc18