[roll] Roll fuchsia [rtl8111] Rewrite in C++
The motivation for this was to migrate FIDL from Banjo to PCI, which
requires the use of a C++ FIDL client. When I changed the file extension
from .c to .cc I found that extensive changes were needed to even get it
to compile (mainly because of gotos), so I went ahead and rewrote it in
idiomatic C++.

Disclaimer: We are not sure what hardware this driver runs on, and
consequently this CL has not been tested.

Original-Bug: 99391
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727017
Original-Revision: cc34d6c2615a1c6d140dc8002acca80f266c2599
GitOrigin-RevId: 1032d632b719e3a7c2c8576355eeec8d546dfa34
Change-Id: Ia8d13d2c0293880cc52f257211cbdf9b12532f58