[roll] Roll fuchsia [image_format] Round up width for caller.
Existing callers sometimes pass a width value that hasn't yet been
rounded up to the next coded_width_divisor alignment, so instead of
returning failure in that case, round up the width for the caller.

If the caller is a producer, the caller is still expected to indicate
a width to other participants that's divisible by coded_width_divisor.

BUG: 269323365
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807987
Original-Revision: a2459c458eb19b9b38deb666a5efdc54751b42e3
GitOrigin-RevId: d2ee3da53b3006d31fe83c83d45bf525e0175f1b
Change-Id: Ia03477fa3548675e16649be24407cb8d62be3f96