[roll] Roll fuchsia [platform-device] Always offer FIDL protocol
Always offer the FIDL protocol. This was not possible to do back when
setting an outgoing directory required the device to isolate.

This should make it easier to migrate all clients to using platform
device FIDL.

Update the `if` statements to a `switch` statement to ensure we check
all variants of the enum.

Original-Bug: 110796
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807228
Original-Revision: 72ca5276468275d52810326d63bc4a008876abd7
GitOrigin-RevId: 5e10189d3177d288c0faa0a2cd87c50fd5d5b996
Change-Id: Ic702f4136423dc855b7e0b915bc0fed53dc99a31