[roll] Roll fuchsia [services] Ignore NOT_DIRECTORY when opening service source directories
FilteredServiceProvider used the flags passed to Open by the client
as-is to open the source service directory. This failed when the flags
had NOT_DIRECTORY, e.g. when connecting directly to a protocol
within the service.

The provider now ignores the NOT_DIRECTORY flag if given by the client,
and always opens the service directory with the DIRECTORY flag.

Original-Bug: 120673
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/796602
Original-Revision: 649ca052af93cd3b11af31a73675974af9720a0b
GitOrigin-RevId: ac818999ac7c66ff60317970555fea8762a5080f
Change-Id: Id7d16c559fac6747e2845086795cc3ea88b14f22