[roll] Roll fuchsia [netstack3] Introduce weakly referenced device IDs
Weak device IDs are IDs that may be held by things that do not make any
claim about the live-ness of a device. For example, sockets may hold a
weak reference to the device ID at rest since sockets do not claim that
the device will be alive for the lifetime of the socket. Instead,
sockets only need to ensure that a device is alive when attempting to
send packets.

Original-Bug: 100759
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807730
Original-Revision: ba3243aaa3a7657062813ac7608bc7a4b999af10
GitOrigin-RevId: 1a42b0fa8f5e126699c974aef6c6510d6c9d2d64
Change-Id: I17937edbfeb399bf26a6c5b82766b7f239486ae0