[roll] Roll fuchsia [ramdevice_client] Add method to wait for removal
Add a method to the rust RamdiskClient that unbinds the ramdisk and
waits for the device and all of its descendants to be removed.

Only calling `schedule_unbind` in `destroy` leaves a small window where
the child block device can still be used after `destroy` returns. The
storage stress tests want to wait for that window to be over so
filesystems can be shutdown without being able to flush anything. The
current method of waiting for the directory connection to be closed
doesn't work because directory connections are not closed when a
directory is unlinked.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/808891
Original-Revision: 772123b6c415d275b470c24c4bc7416caa0fdcfd
GitOrigin-RevId: 7018862d2145f69fe5024509b04bef69ebb7a141
Change-Id: I4acbd5ac6b825e962b41520619cbe6e94b98dc3b