[roll] Roll fuchsia [trace] Enable Tracing via common_minimum_eng AIB
When building eng builds, trace manager is added as a child to core so
that trace providers can connect to it. This allows us to remove
tracing in non eng builds by excluding it in the core shard as opposed
to the current behaviour which is to always including it in the core
shard as a child and simply not provide the trace_manager package,
which works, but causes verbose logs spew when the package is requested
but not found.

This is a stepping stone -- this cl fixes enough optional routes that
building without the core shard is only works for non-eng core right
now. Non-eng Workstation needs additional work. We'll need to do a
through cleaning of the fuchsia.tracing.provider.Registry capability
routing to removing usages of it that are non optional.

Original-Bug: 112433
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779082
Size-Review: Gopi Krishna Chitluri <gopichitluri@google.com>
Original-Revision: 4999eba6e7d804698f7fd26a6b5d89883aa0b1ae
GitOrigin-RevId: 80c8e15e03165ae6bef9e0d3fce9f2e769f20767
Change-Id: Ia2260d3e23adacd92e0f0a31927a2cbeb18e4d91