[roll] Roll fuchsia [pistress] Update pistress to include NO_INHERIT profiles.
Update the PI stress test to also include fair profile threads which
do not have inheritable profile.

Also, stop passing periods which are not equal to deadlines.  Period
was effectively removed from the deadline scheduler parameters, and
should eventually be removed from the profile interface, so there is
not much point in us passing it.

Finally, add pistress to the zircon debugging bundle so that it will
get built by the bots, and hopefully will not end up breaking due to
changes in other parts of the system.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793566
Original-Revision: 2929cba11711c30ff159cce4e670ba70e5a394c1
GitOrigin-RevId: a290a6c9bf423948895db977f9a7fd1c3fcd5401
Change-Id: I8ef6782cd4f911d118ec40c3fdebb3eb7cacb160