[roll] Roll fuchsia [inspect][e2e] Create system-wide inspect benchmark
This change contributes system-wide benchmark for Inspect that reports
the behavior of reading Inspect data from each component on the system.

Further work will be required to wire up the fuchsiaperf files output by
this test with the processing pipelines, but that will be part of a
follow-up.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803254
Original-Revision: 03f0d02e27012505cbec4bc2ecaf854de87547d2
GitOrigin-RevId: f2a3eee49abb932198baac0ddda5ae1e6f16b760
Change-Id: I8e6f751dff1397117017c79fdf346fb738ea7249