[roll] Roll fuchsia [kernel] Remove bad dependency in validate-kernel-symbols rule
The action uses the link_output_rspfile() target in deps to
deduce the input file and gets it into the depfile, so this
`inputs` list is superfluous and erroneous in some variant
configurations. However, that's only true when the script
is fixed to actually emit the right things in its depfile.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809528
Original-Revision: f20851a3d2cfadab185416187fe5c25cd0f49956
GitOrigin-RevId: 05c6cd7be82dd1d4b5b7be6bc6907221ac7d065f
Change-Id: Iafbba4b11355997c6ad07e34903e0fc5612b231d