[roll] Roll fuchsia [netstack] Add network-with-tracing-basic
Currently the network realm that has tracing enabled is configured to
include advanced features, which is the default. To make it easier to
trace on products that require the basic network realm, add a separate
variant of the network package for that and update the documentation to
explain it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810047
Original-Revision: e4121edc4b03724acc2dd9c7a6f97b4b5f9b5de4
GitOrigin-RevId: 4d1c9b27d33d40efa6ba810696dad2d636f6007c
Change-Id: Ic2bf159abd5ac9221be9ed47388e17ea65dd8464