[roll] Roll fuchsia [products] Remove device-finder from bringup host_labels
Product files should only configure target-side packages. Any inclusion
of host-side targets in a product file is considered a hacky developer
convenience, and will often end up over-building, e.g. in infrastructure
when most tools for local development are not needed, so should be
disallowed.

device-finder was originally added to bringup.gni in
https://fxrev.dev/369050, but is no longer needed by default now that
it's not used by any fx commands (and any fx commands that do require it
should build it on-the-fly instead of assuming that it already exists on
disk).

Original-Bug: 120755
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809988
Original-Revision: f99567932a59fdad0323bc7939c891045eb6cb3c
GitOrigin-RevId: e8620fabc18f659f34ea36cccccf06a5daed2a9a
Change-Id: Ia1ea16b633fdd65201beb45bb3ac80bc31aef727