[roll] Roll fuchsia [phys] Update ElfImage API for clarity (and for a few bugs)
Previously, out-of-process loading was expected to use ElfImage::Load()
followed by ElfImage::SetLoadAddress() giving the actually desired
virtual address. This led to a bug in which Load() assumes that the
final load address is to be where the image ends up in the current
address space: it tells gSymbolize about this, which prints the wrong
load address. Accordingly, we update ElfImage::Load() to optionally take
a load address up front.

Out-of-process loading also needs the address of the memory image in the
current address space, however, which is abstracted away by the current
interface. Previously, the only way to access this would be to call
`LoadAddress()` after `Load()`, but before `SetLoadAddress()` to specify
the address under the new mapping: this is awkward - and also is no
longer feasible with the change described above. Accordingly, we now
provide an accessor for the image in the current address space as well.

Further `elfldltl::LoadInfo load()` has been renamed to `load_info()` to
disambiguate it from `Load()`.

Original-Bug: 91187
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/798902
Original-Revision: 99d4acded6fb73882e51028b2d3e9a887caaa2f2
GitOrigin-RevId: 189abd26f76573faf3bbeb6453e8b8fc879bc759
Change-Id: I4847c088ff272bcdf85093e493f016adfb78c247