[roll] Roll fuchsia [starnix] Get binder selinux context from SeLinuxThreadGroupState
The task now keep its current selinux context in
SeLinuxThreadGroupState. This CL updates the binder driver to retrieve
the context from there instead of hardcoding a fake one.

This is still incorrect for remote binder that are not associated to a
real task.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810488
Original-Revision: 693cf9a5045221206392dca01be4988c019cd159
GitOrigin-RevId: 7b915a6e6ec5599edc7e562a65832210f8fab03a
Change-Id: Ib0e74438bcac20116def8d64f9e2756a6c7f65c0