[roll] Roll fuchsia [events] Disambiguate between multiple "use" statements with the same name"
In order to fix a bug which causes a component to not be able to listen to
more than one CapabilityRequested events.
This CL modifies routing so that it takes a Use decl or
a name. In the case of a use decl, it compares against the entire
decl when performing routing, rather than just the name.
Note that tools such as scrutiny and the routing lib tests
still use names rather than use decls.

Run-All-Tests: true

Original-Bug: 121625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800163
Original-Revision: 95c36e5157528daefaf0584f39982edec858905d
GitOrigin-RevId: 188a63781a5f065528d85534f81182b0a454361e
Change-Id: Ie0978e69d85561b0318fdb8a43eb4abc2741620a