[roll] Roll fuchsia [media][codec] Teardown loop before CodecFactoryApp
If CreateChild or OpenExposedDir callbacks in ForwardToIsolate are outstanding when CodecFactoryApp is destroyed, then the failure_func callback will remain alive after the CodecFactoryApp. When the loop is shutdown, that will cause the FIDL bindings to get errors and failure_func to be called. For CodecFactoryApp::DiscoverMagmaCodecDriversAndListenForMoreAsync, this will cause TeardownMagmaCodec to be called, which will cause a UAF of the CodecFactoryApp.

To prevent this, always shutdown the loop before destrying the
CodecFactoryApp.

Original-Bug: 121693
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/801427
Original-Revision: 37b6baf33b803a90419d28cc34507124cd0fe5a1
GitOrigin-RevId: fb50aeabfadfe0002f44bb93de8e71c7f7be2c7d
Change-Id: I1c759f191de673975ee793757ec5a3cc2e1cc704