[roll] Roll fuchsia [kernel] Centralize startup failure detection.
Also fix a race in the process.

There currently exist a few different places where the kernel attempts
to ensure that all CPUs have properly started before transitioning to
usermode.

One of these sync points was added a long time ago in two CLs:
+ fxr/425312
+ fxr/425313

While the mechanism to wait-for-sync added was generic, it was only
used in one specific place on ARM64, in order to wait to figure out
which version of `zx_ticks_get` to install in the VDSO (all CPUs
needed to have started in order to check to see if any suffered from a
timer errata which needed to be mitigated in software).

x64 already had a much shorter blocking timeout for waiting for
processors to boot, and a new non-blocking check was added to ARM in
this CL, fxr/743503.

This CL introduced a bit of trouble, however.  It created a thread
which would sleep for a fixed 5 seconds, and then check to make sure
that all CPUs had reported themselves as being online, printing an
OOPS to be picked up by CI/CQ if and of them hadn't.  While this is
almost always fine, there is a possibility that a CPU is taken offline
after having successful come online, but before the check runs.

In particular, when we run mexec boot tests, the system boot into
usermode, and then almost immediately mexecs, which takes all CPUs
offline before jumping to the mexec image.  So, if we hit a specific
timing window, where we have started to shutdown CPUs, but have not
actually shut them all down, and the timer for this check fires, it
can end up reporting that a CPU never came online, even though it did.

This CL makes and attempt to unify and centralize all of this
behavior, and to eliminate the race in the process.  It changes the
check gating the launching of usermode to always be run, not just on
ARM64.  The ARM64 specific timer-mitigation code is still in place, it
just does not have to wait anymore (since we now wait for CPU sync a
bit earlier on).  If not all of the CPUs have sync'ed up, we print an
OOPS, and then go ahead and print an additional OOPS for any CPU which
is not marked as online.  This should help to disambiguate between a
situation where the CPU just never came online, or wedged _very_ early
in the startup sequence, vs a CPU which made it farther into its
bringup, but still managed to wedge later on.

In addition, this makes the sync-up point common for all
architectures, x86/64, ARM64, and anything added in the future.

With the new checks in place, we can remove the non-blocking thread
check which was added and was leading to the race.

Original-Bug: 121832
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/809227
Original-Revision: 79b23f61071b470934e6ea138fac9f8a030eed12
GitOrigin-RevId: edaebd8b845d2d0351bdb858db9d8ab0f27bca1d
Change-Id: I5cee08a43ef1efd369ef4b499c95bccd9eb37e2f