[roll] Roll llvm-project/libc [libc] use vars in string to num fuzz targets
The string to integer and string to float standalone fuzz targets just
ran the functions and didn't do anything with the output. This was
intentional, since they are intended to be used with sanitizers to
detect buffer overflow bugs. Not using the variables was causing compile
warnings, so this patch adds trivial checks to use the variables.

Reviewed By: sivachandra, lntue

Differential Revision: https://reviews.llvm.org/D144208

GitOrigin-RevId: 4ce707d3853e815826e376a87d74917440de8222
Original-Revision: 6dfc7e95977193d2cb44ee30f1cf491982c2b543
Change-Id: I072ca034db12e796d0721220a9ba38e71cf4c239