[roll] Roll fuchsia [ffx test] Remove suite_start_timeout_seconds timeout
This timeout was added to address hangs due to ffx stability bugs. If
ffx lost connection to the device, the test used to hang. This timeout
has unintended consequences, as on device processing can take long time
for large packages and is unreliable. So remove the timeout.

In future if we see ffx stability issues, we should add a heartbeat API
to test manager and periodically use that to timeout on connection loss,
or overnet should do it internally.

BUG=122105
BUG=122144

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/805889
Original-Revision: e22b8ae0c38339a1536d0db4a346ebd97d8f897f
GitOrigin-RevId: 9b6c168ee49116c37b28cd16efb206060acb68e4
Change-Id: If813e2d7176f4f8e3f6a08087722e62545d6b19f