[roll] Roll fuchsia [network] Add an AIB for the netstack package
Move the selection of the netstack package out of the legacy mechanism
for the minimal product and into product assembly. This will enable the
choice of netstack to be configurable at assembly time, though for now
it is a no-op.

Original-Bug: 85450, 121811
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/807750
Original-Revision: 2c9c66de1c4373f604dc1ec49d4b00976358f25a
GitOrigin-RevId: b708e02a68c1bd514b04658575030d5f82f84c7c
Change-Id: Ieaf1b5e51c783112b177eb64ac4fb4eae9ab0293