[roll] Roll infra/recipes [recipes] Set riscv64 as the target when TCG is enabled
Unfortunately, the build graph does not correctly set the target_cpu to
x64 when building the riscv64 phys tests. We could add another boolean
to communicate this to the recipe that constructs the qemu config
(testing_requests/api.py), but that seems like a lot of overhead given
that the build will eventually need to correctly set this bit.
Therefore, use a far more hacky approach in which we assume that any
builder that specifies `use_tcg` will be running RISC-V targets (which
is true for now).

Original-Bug: b/270961878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/811514
Original-Revision: 0f919e6063a2e307430a038da0cf43ed7f90df62
GitOrigin-RevId: 67c9abbf537f3b28a782c5e0a0ff3f26b7950b9f
Change-Id: I7a461f2e0a0d28bd9ac7b1e729e1b22f5c38d6c0