[roll] Roll fuchsia [netstack3] Choose netstack via assembly config
Add a configuration arg to choose which of Netstack2 and Netstack3 is
used in the product image at product assembly time based on platform
config or board feature.

Add a new "core-with-netstack3" product to be used in CI builds and
temporarily propagate use of the "use_netstack3" GN arg to allow
overriding the choice of netstack for the existing "core" product. The
latter portion will be reverted once the Netstack3 CI builders are
migrated to use the new product.

Original-Bug: 121811
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810887
Original-Revision: b5dde4dcb9452aa5e2da5deec07fd0ef84846392
GitOrigin-RevId: 20af5d769c88d7d5e4a2b80afc1d8f8d00b0e2af
Change-Id: Ie6930fedcf6402966c6d29bdce0a6a4accbe71a1