[roll] Roll fuchsia [netstack3] Store `user_timeout_until` for TCP retrans timers
... instead of `start`. As pointed out in this [comment], this is more
consistent with other protocols in netstack3, so that changing a socket
option doesn't take effect immediately but rather at the next time a
timer needs to be created. This is not exactly what Linux does, but it
is also unlikely that an application can depend on this and tell
difference.

[comment]: https://fuchsia-review.git.corp.google.com/c/fuchsia/+/808350/comment/4d3703e4_4ebb2ce2/

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810598
Original-Revision: 2a143d5dca01e1051bcc70e84b7094264b5a0286
GitOrigin-RevId: 8a1a67950615051470dad2787631655ed7be4944
Change-Id: Ic1bb133d214b913efa70356841e1a43229355074