[roll] Roll fuchsia [pixel][test] Robustifies pixel tests
The tests as written were sensitive to the exact
nuance of colors used to examine the scene
histogram.

The new code matches the colors approximately, and does not
rely on exact color matching. Not quite a SNR calculation,
but enough for this test.

I also added a histogram printout as part of the test logs,
which hopefully offers more insight into why the tests might
fail if (when) they do.

Original-Bug: 112452
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811967
Original-Revision: d164e01ff368a27939c6a27636103c72899d4c78
GitOrigin-RevId: 5c731b1975542abe0406ec9ea8b8c69a5728fb9d
Change-Id: Ic5acd7b5ab911f97ca5537378c30e3fa7a57c841