[roll] Roll fuchsia [async_patterns] Bridge to std::future<R>
When using [Test]DispatcherBound in driver tests, we discovered that
SyncCall does not work if the test is in the "driver runtime does not
manage thread" mode. In that mode, the user needs to manually run the
dispatchers using fdf_run_until_idle. AsyncCall works but is very
verbose, requiring setting up a Receiver object. This CL adds a
ToFuture function that turns the result of an async call into a
`std::future` which can be awaited in tests synchronously.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811847
Original-Revision: 5c741840deeaeea49e30550268e100e3c48f58c1
GitOrigin-RevId: 475fa48f50330f1ad7e3684d956fc82f5cec1423
Change-Id: I5c44f3e3d9201436420e140ae3cb0c82b941b80b