[roll] Roll fuchsia [fdf] Wait for DFv2 node to be removed completely
This CL addresses two issues related to shutting down DFv2 nodes. The
first issue is that DFv2 nodes do not confirm that its associated
component has been fully removed by the component manager before
allowing another component to be created with the same configuration. If
a new node is added after the old one is removed and the component
manager hasn't had time to remove the old component then it may complain
about duplicate components when the new node attempts to create a new
component.

In this CL, nodes will now send a `DestroyChild` request to the
component manager during the node removal process and will only be
considered fully removed when the FIDL request gets a response. This
ensures that the associated component has been fully removed.

The second issue is that DFv2 nodes can be added after another node with
the same name has been requested to be removed but before it has been
fully removed. The driver runner will complain that duplicate nodes
exist when attempting to add the new node even if the old node is in the
middle of being removed.

In this CL, DFv2 nodes will now check if the duplicate node is in the
middle of being removed and create a callback to create the new child
node once the old node is removed.

Both these issues are race conditions and so the test added in this CL
may not fail every time. It usually failed every 1 out of 3 tests.

Multiply: rebind_test
Original-Bug: 121991
Original-Bug: 121990
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/800764
Original-Revision: de7e754be1ae0886a590a890eb29a97165b18f60
GitOrigin-RevId: 66e52b5a67a84028034346d148fecf64188019ad
Change-Id: Ia4882065b0cf49e5ec4a6cb224c89199718329bb