[roll] Roll infra/recipes [recipes] Override emulator memory to 8GB for RISC-V
The RISC-V boot shim hardcodes a 8GB memory range, so the emulator must
be started with that much memory. This should likely be a part of the
test spec in the long term, but add a workaround here to unblock the
program.

Original-Bug: b/270961878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/812487
Original-Revision: 6c2e05a662dfe20a3fc03290a427f422b1587ba5
GitOrigin-RevId: 2d2f861cec60c24cf0e22b6b00244c0b63f65b63
Change-Id: If2b01407580e10e5e4c11a4b1d861573630f936f