[roll] Roll fuchsia [kbma] Introduce counter increment/decrement logic.
This CL introduces the hooks through which attribution objects
are incremented and decremented. The instrumentation through
which cow pages are given an attribution object will be introduced
in a follow up cl, so for now even with kbma enabled this cl
would be a no-op.

Original-Bug: 117196

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/802267
Original-Revision: f87f64758c14c18317b34df4b26352e9f3db9cff
GitOrigin-RevId: 2bc6cb1f42a89913b639d3e00a1cdcb033b6e845
Change-Id: Iab9833f56e033d58bbc4e1d899a53b8caa7e47fb