[roll] Roll fuchsia [time] Register the pull time source
The pull time source is registered as a new component
in the product assembly, and the core size limits
calculations.

I expect little effect on the size of the time bundle,
as the component is identical to the push component,
and the config files should be small.

Original-Bug: 113956
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803368
Size-Review: Aaron Wood <aaronwood@google.com>
Original-Revision: d6b2154029253897565da97959663c92a53f13d7
GitOrigin-RevId: dbb3edbd6a23fce2c8faba9e1cbb176939a66834
Change-Id: Ia6c80f050abca80e0c8ab29f427d099d0e2e6a83