[roll] Roll fuchsia [botanist] Get QEMU kernel using image name and CPU
The kernel that botanist passed to QEMU was previously just the first
image with type `qemu-kernel` in images.json. This is a problem when
multiple QEMU kernels are present with different target architectures
(as in the case of our RISC-V builder). This CL solves the problem by
filtering by both the image name and the CPU of the associated image.

Original-Bug: b/270961878
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812849
Original-Revision: 6eba5eebdf95c2c3562fc057639aba93fb129859
GitOrigin-RevId: b70030327e757134bb99d1e0a7bbfe936f9aa7ba
Change-Id: I3a7d8985c028bbc2c9139afa5c252705e7a37c54