[roll] Roll fuchsia [netstack3] Get Buffer len and capacity atomically
Without this, a future change to make the RingBuffer held in a
SendBufferWithZirconSocket resizable would introduce raciness. Since
both len() and capacity() would depend on the fullness of the zircon
socket, and that can change due to the process attached to the other end
reading, the values returned by len() and capacity() would not be
guaranteed to be consistent with one another. Prevent this by, in this
change, returning both in the result of a single method.

Original-Bug: 115456
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/812987
Original-Revision: e455433ce1c13b0e1e316811ca12d44ee3e6f1c9
GitOrigin-RevId: 17404903f0b18dc0394b7b1a3048758c4a714791
Change-Id: I378d36f891736db163caa6fc101fa02ed11f4510