[roll] Roll fuchsia [botanist] Have CSO builder using CSO-only mode.
The CSO builder is using ffx experiment level 4, and since the rest of
the ffx builders are running with CSO in enabled mode, we can use the
CSO builder to test the CSO-only mode.
Also move enabled mode to experiment level 1 so we can see if it helps
with flashing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813054
Original-Revision: dd49d42cd024ed2e849965c8fa95b6462376aee9
GitOrigin-RevId: 87c36ca480b25b5be8f72d369b651293a041c7e2
Change-Id: Iaeb4d4b60040c018300992c906f83123f22a2bbd