[roll] Roll fuchsia [msd-arm-mali] Process job interrupts using port
Job interrupts were waited on by the job interrupt thread, which posted
a request to the device thread and waited for a reply. The extra work
takes at least 25 us of CPU time per interrupt (on astro), which adds up
over the course of 10-20 interrupts for frame.

Instead we should bind the job interrupt to a port, so the device thread
can wait on it directly.

Original-Bug: 122629
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811809
Original-Revision: 3b3a15a35e911f7406b9aaf3ef79bb6cc0169095
GitOrigin-RevId: afbad099d06b66377d3a1865e222fbc29e2b5afe
Change-Id: I712bc113c84f07b5317e24fa8875383cc76f4d9c