[roll] Roll fuchsia [gn_desc] Optimizations to increase startup speed by 60 percent
Changes the optimization level to 'speed' for the binary and it's rlibs
as well as reading the entire json file into a String buffer before it
starts to parse it with serde.

Half the improvement is from reading into a String, the other half from
the increased optimization level.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813368
Original-Revision: 0a96d4ed14e56826b38c7e23efeb51fe50cff5ef
GitOrigin-RevId: 939fb39c7e213f42d6a227b796619c145b8490ac
Change-Id: I5824c5e99ff3152f8655d8c5cd544a3c820db8e8