[roll] Roll fuchsia [uart] Introduce SyncPolicy
A SyncPolicy provides the mechanisms required for
synchronization. That is:
    * Lock Type: Opaque capability
    * Guard Type: Opaque RAII for acquiring the lock
    * Waiter Type: Mechanism for waiting on
                   Tx Hw Fifo becoming non full.
    * AssertHeld: Assert a Lock instance is held.
    * DefaultLockPolicy: Default Lock Policy for methods.

The Guard type is allowed to have one template argument, which is called
LockPolicy. A LockPolicy is defined on the driver methods, such that
the same object can use different flavors. This allows for the
kernel environment to implement:
    * IrqSave
    * NoIrqSave
    * NoopLocking (Panic Shell)
See following patch for details.

Original-Bug: 89182
Test: fx test uart-unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/803047
Original-Revision: 3cc670d3ae655f2f6bae0fc2816eaf61de8ac9fd
GitOrigin-RevId: c3591fbf24c8c4f21f5affebea6a1ee0bbf1934c
Change-Id: Id4165a6cda79280f72d237092d1d9da7d8af8ddf